+++ web/src/main/java/org/geogebra/web/geogebra3D/web/euclidian3D/EuclidianController3DW.java
@@ -232,7 +232,7 @@ public class EuclidianController3DW extends EuclidianController3D implements
 	 * @return the multiplier that must be used to multiply the native event
 	 *         coordinates
 	 */
-	public float getScaleXMultiplier() {
+	public double getScaleXMultiplier() {
 		return mtg.getScaleXMultiplier();
 	}
 
@@ -240,7 +240,7 @@ public class EuclidianController3DW extends EuclidianController3D implements
 	 * @return the multiplier that must be used to multiply the native event
 	 *         coordinates
 	 */
-	public float getScaleYMultiplier() {
+	public double getScaleYMultiplier() {
 		return mtg.getScaleYMultiplier();
 	}
